Sync of Windows/ESENT code to Exchange/ESE
Move Windows/ESENT commits to Exchange/ESE through OS commit 489831f1c149a

[Substrate:a604a4fb00044250d35a990bf3d6a0d67566b400]

+++ dev/ese/src/ese/_log/logredo.cxx
@@ -2304,7 +2304,7 @@ LOCAL ERR ErrLGRIMacroClearRedoMapDbtimeRevert( PIB* const ppib, const DBTIME db
     {
         size_t  cfmppgnofreed = prgfmppgnoFreed->Size();
 
-        for ( int i = 0; i < cfmppgnofreed; ++i )
+        for ( size_t i = 0; i < cfmppgnofreed; ++i )
         {
             CFMPPage fmppage = prgfmppgnoFreed->Entry( i );
